feat(ts): reorg. TODO endianness bugs?
