+++ ts/src/utils/jsUtils.ts
@@ -1,4 +1,10 @@
-export function range(n: number, offset: number = 0) {
+/**
+ * Create an array over an integer range.
+ * @param n - The number of integers.
+ * @param offset - The starting number.
+ * @returns The array of numbers.
+ */
+export function range(n: number, offset = 0) {
   const ret: number[] = [];
   for (let i = 0; i < n; i++) {
     ret.push(offset + i);
@@ -6,6 +12,12 @@ export function range(n: number, offset: number = 0) {
   return ret;
 }
 
+/**
+ * Assert a member is a certain length.
+ * @param obj - An object.
+ * @param member - A member string.
+ * @param length - The length.
+ */
 export function assertLength(obj: any, member: string, length: number) {
   if (obj[member].length !== length) {
     throw new Error(
@@ -13,3 +25,11 @@ export function assertLength(obj: any, member: string, length: number) {
     );
   }
 }
+
+/**
+ * Strips methods of a type.
+ */
+export type FieldsOf<T> = {
+  // eslint-disable-next-line @typescript-eslint/ban-types
+  [P in keyof T as T[P] extends Function ? never : P]: T[P];
+};
